requirements.txt
folium
geopandas
networkx
numpy
ortools
osmnx
pandas
pyarrow
pytest
python-dotenv
requests
shapely
